requirements.txt
numpy~=1.21.4
Pillow~=8.4.0